test: Update missed tests for rv64
This change updates tests that were missed in previous PR to support
RV64.

TEST=Verified that all tests pass for rv64 now using
tests-rv64-qemu-virt suite.

+++ test/test-assist/src/main.rs
@@ -133,7 +133,7 @@ fn illaccess(_arg: u32) {
 #[inline(never)]
 fn illfunc(_arg: u32) {
     unsafe {
-        let f: extern "C" fn() = core::mem::transmute(_arg);
+        let f: extern "C" fn() = core::mem::transmute(_arg as usize);
         f();
     }
 }
